Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -4,7 +4,7 @@ int main() {
     float d;
     scanf("%d %d %d",&a,&b,&c);
     d=a+b+c;
-    int avg=d/3;
+    float avg=d/3;
     printf("Average: %.2f",avg);
     return 0;
 }
